Switch LINKER_LANGUAGE from C to CXX
On compiling postgresql and lanterndb from source from scratch, without
installing postgresql core, the lanterndb extension compiles and
installs, but it is unable to execute `CREATE EXTENSION lanterndb;` with
the error `undefined symbol: _ZTVN10__cxxabiv117__class_type_infoE`.
This is due to libstc++ not being linked. Switching the LINKER_LANGUAGE
to CXX allows the build system to use a C++ linker and solves the issue.

References:
- https://github.com/Iorethan/opencv_pfm/issues/1
-